+++ src/cli/subcommands/snapshot_cmd.rs
@@ -6,6 +6,7 @@ use crate::chain_sync::chain_muxer::DEFAULT_RECENT_STATE_ROOTS;
 use crate::cli_shared::snapshot::{self, TrustedVendor};
 use crate::db::car::forest::new_forest_car_temp_path_in;
 use crate::networks::calibnet;
+use crate::rpc::chain::ForestChainExportDiffParams;
 use crate::rpc::{self, chain::ForestChainExportParams, prelude::*, types::ApiTipsetKey};
 use anyhow::Context as _;
 use chrono::DateTime;
@@ -39,6 +40,21 @@ pub enum SnapshotCommands {
         /// Snapshot format to export.
         #[arg(long, value_enum, default_value_t = FilecoinSnapshotVersion::V1)]
         format: FilecoinSnapshotVersion,
+    },
+    /// Export a diff snapshot between `from` and `to` epochs to `<output_path>`
+    ExportDiff {
+        /// `./forest_snapshot_diff_{chain}_{from}_{to}+{depth}.car.zst`.
+        #[arg(short, long, default_value = ".", verbatim_doc_comment)]
+        output_path: PathBuf,
+        /// Epoch to export from
+        #[arg(long)]
+        from: i64,
+        /// Epoch to diff against
+        #[arg(long)]
+        to: i64,
+        /// How many state-roots to include. Lower limit is 900 for `calibnet` and `mainnet`.
+        #[arg(short, long)]
+        depth: Option<crate::chain::ChainEpochDelta>,
     },
 }
 
@@ -140,6 +156,77 @@ impl SnapshotCommands {
                     temp_path.persist(output_path)?;
                 }
 
+                println!("Export completed.");
+                Ok(())
+            }
+            Self::ExportDiff {
+                output_path,
+                from,
+                to,
+                depth,
+            } => {
+                let raw_network_name = StateNetworkName::call(&client, ()).await?;
+
+                // For historical reasons and backwards compatibility if snapshot services or their
+                // consumers relied on the `calibnet`, we use `calibnet` as the chain name.
+                let chain_name = if raw_network_name == calibnet::NETWORK_GENESIS_NAME {
+                    calibnet::NETWORK_COMMON_NAME
+                } else {
+                    raw_network_name.as_str()
+                };
+
+                let depth = depth.unwrap_or_else(|| from - to);
+                anyhow::ensure!(depth > 0, "depth must be positive");
+
+                let output_path = match output_path.is_dir() {
+                    true => output_path.join(format!(
+                        "forest_snapshot_diff_{chain_name}_{from}_{to}+{depth}.car.zst"
+                    )),
+                    false => output_path.clone(),
+                };
+
+                let output_dir = output_path.parent().context("invalid output path")?;
+                let temp_path = new_forest_car_temp_path_in(output_dir)?;
+
+                let params = ForestChainExportDiffParams {
+                    output_path: temp_path.to_path_buf(),
+                    from,
+                    to,
+                    depth,
+                };
+
+                let pb = ProgressBar::new_spinner().with_style(
+                    ProgressStyle::with_template(
+                        "{spinner} {msg} {binary_total_bytes} written in {elapsed} ({binary_bytes_per_sec})",
+                    )
+                    .expect("indicatif template must be valid"),
+                ).with_message(format!("Exporting {} ...", output_path.display()));
+                pb.enable_steady_tick(std::time::Duration::from_millis(80));
+                let handle = tokio::spawn({
+                    let path: PathBuf = (&temp_path).into();
+                    let pb = pb.clone();
+                    let mut interval = tokio::time::interval(tokio::time::Duration::from_secs(1));
+                    async move {
+                        loop {
+                            interval.tick().await;
+                            if let Ok(meta) = std::fs::metadata(&path) {
+                                pb.set_position(meta.len());
+                            }
+                        }
+                    }
+                });
+
+                // Manually construct RpcRequest because snapshot export could
+                // take a few hours on mainnet
+                client
+                    .call(ForestChainExportDiff::request((params,))?.with_timeout(Duration::MAX))
+                    .await?;
+
+                handle.abort();
+                pb.finish();
+                _ = handle.await;
+
+                temp_path.persist(output_path)?;
                 println!("Export completed.");
                 Ok(())
             }